[main] Ajustar: entidad (prop alta), propiedades (especificaciones).

+++ namasdev.Apps/namasdev.Apps.Entidades/PropiedadTipoEspecificacionesNumeroLargo.cs
@@ -1,6 +1,6 @@
 ﻿namespace namasdev.Apps.Entidades
 {
-    public class PropiedadTipoEspecificacionesNumeroLargo : PropiedadTipoEspecificacionesNumeroBase<double>
+    public class PropiedadTipoEspecificacionesNumeroLargo : PropiedadTipoEspecificacionesNumeroBase<decimal>
     {
     }
 }
